Rest: Move Dialer to its own file
